+++ mathutil/mathutil.go
@@ -19,6 +19,10 @@ type Integer interface {
 		~uint | ~uint8 | ~uint16 | ~uint32 | ~uint64
 }
 
+type Float interface {
+	~float32 | ~float64
+}
+
 type Numeric interface {
 	~int | ~int8 | ~int16 | ~int32 | ~int64 |
 		~uint | ~uint8 | ~uint16 | ~uint32 | ~uint64 | ~uintptr |