+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### [13.12.0](https://kaos.sh/ek/13.12.0)
 
 * `[req]` Added custom timeout per request
+* `[req]` Added `Retrier`
 * `[req]` Make `Limiter` public
 * `[log]` Added `WithFullCallerPath` option to enable the output of the full caller path