+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * `[log]` Added `WithFullCallerPath` option to enable the output of the full caller path
 * `[strutil]` Added support of escaped strings to `Fields`
 * `[strutil]` Added fuzz tests for `Fields` method
+* `[knf]` Fixed build of fuzz tests
 
 ### [13.11.0](https://kaos.sh/ek/13.11.0)